airflow_opa/requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.9
python-binance==1.0.19
python-dateutil==2.9.0.post0
awscli==1.33.21
boto3==1.34.139